Initial code
The initial goal of this code is to access all the data files from both the data folders. On running the code on the Jupyter notebook (python 3.11), it displays latitude vs longitude graphs and altitude vs time graph. Several improvements and implementation are to be made for cleaning the data/preprocessing the data, building, training and testing the data model.