Implemented Random Forest Regression model
Random Forest Regression model was implemented with the output.csv file which had a short dataset. Conducted cross validation test between DecisionTreeClassifier, RandomForestClassifier, LinearRegression and RandomForestRegressor models and found that RandomForestRegressor model has a good accuracy score. On the other hand, the Linear Regression model accuracy was quite high and led me to believe the data is overfitting due to less volume. Next steps are to create some noise in the data and then in the meanwhile, we can collect new data everyday.