Implemented 3-layered feed forward neural network
This is different than the random forest regression model since it only calculated the hours and not the minutes. With the feed forward neural network, the minutes are also being calculated. There are 1440 minutes in a day and according to that the predicted ATA is being compared to the ETA.