Model Completed nn.ipynb
The neural network model is completed and there was a slight modification to the features that were being input into the neural network. Also, sample data can be put inside the neural network and it will give out the prediction of the actual time of arrival. 
The error rate is high, so the next task is to focus on that.